+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-<!-- The next build on `develop` should refer to version 1.4.1 -->
+<!-- The next build on `develop` should refer to version 1.5.0 -->
 
 ## [1.4.0] - 2023-05-12
 ### Fixed